Simplify mock data setup
