+++ api/app/tests/weather_models/crud.py
@@ -128,90 +128,3 @@ def get_predicted_daily_precip(*args):
             prediction_timestamp=datetime(2023,10,11,20,0,0)
         )
     ]
-
-
-def get_actuals_and_predictions_with_high_rh(*args):
-    """ Fixed response as replacement for app.db.crud.observations.get_actuals_left_outer_join_with_predictions
-        that fixes actual and predicted rh values at 105 to test clamped values of 100.
-    """
-    result = [
-        # day 1
-        [HourlyActual(
-            weather_date=datetime(2020, 10, 10, 18),
-            temperature=20,
-            temp_valid=True,
-            wind_speed=10,
-            wind_direction=90,
-            relative_humidity=105,
-            rh_valid=True),
-         ModelRunPrediction(
-            tmp_tgl_2=2,
-            rh_tgl_2=105,
-            apcp_sfc_0=2,
-            wind_tgl_10=11,
-            wdir_tgl_10=97,
-            prediction_timestamp=datetime(2020, 10, 10, 18))],
-        [HourlyActual(weather_date=datetime(2020, 10, 10, 19)), None],
-        [HourlyActual(weather_date=datetime(2020, 10, 10, 20),
-                      temperature=25,
-                      wind_speed=15,
-                      wind_direction=270,
-                      temp_valid=True,
-                      relative_humidity=105,
-                      rh_valid=True), None],
-        [HourlyActual(
-            weather_date=datetime(2020, 10, 10, 21),
-            temperature=30,
-            temp_valid=True,
-            wind_speed=12,
-            wind_direction=120,
-            relative_humidity=105,
-            rh_valid=True),
-            ModelRunPrediction(
-            tmp_tgl_2=1,
-            rh_tgl_2=105,
-            apcp_sfc_0=3,
-            wind_tgl_10=11,
-            wdir_tgl_10=101,
-            prediction_timestamp=datetime(2020, 10, 10, 21))],
-        # day 2
-        [HourlyActual(
-            weather_date=datetime(2020, 10, 11, 18),
-            temperature=20,
-            temp_valid=True,
-            wind_speed=9,
-            wind_direction=121,
-            relative_humidity=105,
-            rh_valid=True),
-            ModelRunPrediction(
-            tmp_tgl_2=2,
-            rh_tgl_2=105,
-            apcp_sfc_0=2,
-            wind_tgl_10=11,
-            wdir_tgl_10=110,
-            prediction_timestamp=datetime(2020, 10, 11, 18))],
-        [HourlyActual(weather_date=datetime(2020, 10, 11, 19)), None],
-        [HourlyActual(weather_date=datetime(2020, 10, 11, 20),
-                      temperature=27,
-                      temp_valid=True,
-                      wind_speed=10,
-                      wind_direction=98,
-                      relative_humidity=105,
-                      rh_valid=True), None],
-        [HourlyActual(
-            weather_date=datetime(2020, 10, 11, 21),
-            temperature=30,
-            wind_speed=9,
-            wind_direction=118,
-            temp_valid=True,
-            relative_humidity=105,
-            rh_valid=True),
-            ModelRunPrediction(
-            tmp_tgl_2=1,
-            rh_tgl_2=105,
-            apcp_sfc_0=3,
-            wind_tgl_10=10,
-            wdir_tgl_10=111,
-            prediction_timestamp=datetime(2020, 10, 11, 21))]
-    ]
-    return result
