+++ api/app/tests/weather_models/test_machine_learning.py
@@ -1,11 +1,11 @@
 from datetime import datetime
 import pytest
+from app.db.models.observations import HourlyActual
 from app.weather_models import machine_learning
 from app.tests.weather_models.crud import (get_actuals_left_outer_join_with_predictions,
-                                           get_actuals_and_predictions_with_high_rh,
                                            get_accumulated_precip_by_24h_interval,
                                            get_predicted_daily_precip)
-from app.db.models.weather_models import PredictionModel
+from app.db.models.weather_models import ModelRunPrediction, PredictionModel
 from app.weather_models.machine_learning import StationMachineLearning
 import math
 
@@ -17,17 +17,11 @@ def mock_get_actuals_left_outer_join_with_predictions(monkeypatch):
                         get_actuals_left_outer_join_with_predictions)
 
 @pytest.fixture()
-def mock_get_actuals_and_predictions_high_rh(monkeypatch):
-    monkeypatch.setattr(machine_learning, 'get_actuals_left_outer_join_with_predictions',
-                        get_actuals_and_predictions_with_high_rh)
-
-@pytest.fixture()
 def mock_get_accumulated_precip_by_24h_interval(monkeypatch):
     """ Mock out call to DB returning actual 24 hour precipitation data """
     monkeypatch.setattr(machine_learning, 'get_accumulated_precip_by_24h_interval',
                         get_accumulated_precip_by_24h_interval)
     
-
 @pytest.fixture()
 def mock_get_predicted_daily_precip(monkeypatch):
     """ Mock out call to DB returning modelled/predicted 24 hour precipitation data """
@@ -59,9 +53,42 @@ def test_bias_adjustment_with_samples(mock_get_actuals_left_outer_join_with_pred
     assert math.isclose(wdir_result, 115.51556685719027)
     assert precip_result == 3
 
-def test_bias_adjustment_of_rh_above_100(mock_get_actuals_and_predictions_high_rh,                                         
+def test_bias_adjustment_of_rh_above_100(monkeypatch,                                         
                                          mock_get_accumulated_precip_by_24h_interval,
                                          mock_get_predicted_daily_precip):
+    
+    def get_actuals_and_predictions_with_high_rh(*args):
+        """ Two actual/prediction pairs to force a prediction of rh above 100. """
+        return [
+            [HourlyActual(weather_date=datetime(2020, 10, 10, 18),
+                          temperature=20,
+                          temp_valid=True,
+                          wind_speed=10,
+                          wind_direction=90,
+                          relative_humidity=105,
+                          rh_valid=True),
+            ModelRunPrediction(tmp_tgl_2=2,
+                               rh_tgl_2=105,
+                               apcp_sfc_0=2,
+                               wind_tgl_10=11,
+                               wdir_tgl_10=97,
+                               prediction_timestamp=datetime(2020, 10, 10, 18))],
+            [HourlyActual(weather_date=datetime(2020, 10, 10, 21),
+                          temperature=30,
+                          temp_valid=True,
+                          wind_speed=12,
+                          wind_direction=120,
+                          relative_humidity=105,
+                          rh_valid=True),
+            ModelRunPrediction(tmp_tgl_2=1,
+                               rh_tgl_2=105,
+                               apcp_sfc_0=3,
+                               wind_tgl_10=11,
+                               wdir_tgl_10=101,
+                               prediction_timestamp=datetime(2020, 10, 10, 21))],
+            ]
+    monkeypatch.setattr(machine_learning, 'get_actuals_left_outer_join_with_predictions',
+                        get_actuals_and_predictions_with_high_rh)
     
     predict_date_with_samples = datetime.fromisoformat("2020-09-03T21:14:51.939836+00:00")
